+++ spec/app/app/serializers/alba/inherited/custom_type_user_serializer.rb
@@ -1,0 +1,9 @@
+module Alba
+  module Inherited
+    class CustomTypeUserSerializer < UserSerializer
+      typelizer_config.inheritance_strategy = :inheritance
+
+      typelize id: [:number, optional: true]
+    end
+  end
+end