+++ spec/app/app/serializers/alba/inherited/deep_user_serializer.rb
@@ -1,0 +1,7 @@
+module Alba
+  module Inherited
+    class DeepUserSerializer < ExtendedUserSerializer
+      attributes :updated_at
+    end
+  end
+end