+++ spec/app/app/serializers/alba/inherited/extended_user_serializer.rb
@@ -1,0 +1,12 @@
+module Alba
+  module Inherited
+    class ExtendedUserSerializer < UserSerializer
+      typelizer_config.inheritance_strategy = :inheritance
+
+      typelize :string
+      attribute :full_name do |object|
+        object.username
+      end
+    end
+  end
+end