+++ spec/app/app/serializers/alba/inherited/transform_keys_user_serializer.rb
@@ -1,0 +1,9 @@
+module Alba
+  module Inherited
+    class TransformKeysUserSerializer < UserSerializer
+      typelizer_config.inheritance_strategy = :inheritance
+
+      transform_keys :lower_camel
+    end
+  end
+end