+++ spec/app/app/serializers/ams/inherited/empty_user_serializer.rb
@@ -1,0 +1,7 @@
+module Ams
+  module Inherited
+    class EmptyUserSerializer < UserSerializer
+      typelizer_config.inheritance_strategy = :inheritance
+    end
+  end
+end